requirements.txt
python-telegram-bot[job-queue, rate-limiter]==20.1
openai>=0.27.0
PyYAML==6.0
pymongo==4.3.3
python-dotenv==0.21.0
cryptomus==1.1
jupyter==1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
pandas>=1.5.3
pydub==0.25.1
mixpanel>=4.10.0
mixpanel_py_async==0.3.0
currencies
python-dotenv
git+https://github.com/gkorepanov/whisper-stream.git@2c1270
git+https://github.com/gkorepanov/llm-tools.git@3eff0e
iso639-lang
more-itertools
tenacity
cryptography
tiktoken
levenshtein